+++ Backend/demo/src/main/java/com/example/demo/dtos/ComentarioDTO.java
@@ -1,8 +1,8 @@
 package com.example.demo.dtos;
 
 public class ComentarioDTO {
-    public Long usuarioId;
-    public Long recetaId;
+    public int usuarioId;
+    public int recetaId;
     public Integer puntuacion;
     public String comentario;
 }
