+++ Backend/demo/src/main/java/com/example/demo/dtos/ComentarioRespuestaDTO.java
@@ -4,15 +4,15 @@ package com.example.demo.dtos;
 import java.time.LocalDateTime;
 
 public class ComentarioRespuestaDTO {
-    public Long id;
-    public Long usuarioId;
+    public int id;
+    public int usuarioId;
     public String aliasUsuario;
     public String nombreReceta;
     public String comentario;
     public LocalDateTime fecha;
     public Boolean aprobado;
 
-    public ComentarioRespuestaDTO(Long id, Long usuarioId, String aliasUsuario, String nombreReceta, String comentario, LocalDateTime fecha, Boolean aprobado) {
+    public ComentarioRespuestaDTO(int id, int usuarioId, String aliasUsuario, String nombreReceta, String comentario, LocalDateTime fecha, Boolean aprobado) {
         this.id = id;
         this.usuarioId = usuarioId;
         this.aliasUsuario = aliasUsuario;
