+++ Backend/demo/src/main/java/com/example/demo/dtos/PasoDTO.java
@@ -1,7 +1,7 @@
 package com.example.demo.dtos;
 
 public class PasoDTO {
-    public Long recetaId;
+    public int recetaId;
     public Integer numero;
     public String descripcion;
     public String imagenUrl;
