+++ Backend/demo/src/main/java/com/example/demo/dtos/RecetaIngredienteDTO.java
@@ -1,7 +1,7 @@
 package com.example.demo.dtos;
 
 public class RecetaIngredienteDTO {
-    public Long recetaId;
-    public Long ingredienteId;
+    public int recetaId;
+    public int ingredienteId;
     public String cantidad;
 }
