+++ Backend/demo/src/main/java/com/example/demo/repositories/ComentarioRepository.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.models.Comentario;
 
-public interface ComentarioRepository extends JpaRepository<Comentario, Long> {
-	List<Comentario> findByRecetaIdAndAprobadoTrue(Long recetaId);
-	List<Comentario> findByUsuarioIdAndRecetaId(Long usuarioId, Long recetaId);
+public interface ComentarioRepository extends JpaRepository<Comentario, Integer> {
+	List<Comentario> findByRecetaIdAndAprobadoTrue(int recetaId);
+	List<Comentario> findByUsuarioIdAndRecetaId(int usuarioId, int recetaId);
 
 
 }
