+++ Backend/demo/src/main/java/com/example/demo/repositories/IngredienteRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.models.Ingrediente;
 
-public interface IngredienteRepository extends JpaRepository<Ingrediente, Long> {
+public interface IngredienteRepository extends JpaRepository<Ingrediente, Integer> {
 	Optional<Ingrediente> findByNombreIgnoreCase(String nombre);
 
 }
