+++ Backend/demo/src/main/java/com/example/demo/repositories/RecetaIngredienteRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.models.RecetaIngrediente;
 
-public interface RecetaIngredienteRepository extends JpaRepository<RecetaIngrediente, Long> {
-	List<RecetaIngrediente> findByRecetaId(Long recetaId);
+public interface RecetaIngredienteRepository extends JpaRepository<RecetaIngrediente, Integer> {
+	List<RecetaIngrediente> findByRecetaId(int recetaId);
 
 }
